requirements.txt
Flask==3.0.3
marshmallow==3.23.1
psycopg2-binary==2.9.10